Reorg to hopefully improve logic of code coverage.
