+++ service/src/networking/SignalCommunicator.ts
@@ -1,4 +1,4 @@
-import { WebrtcSignalingRequest, WebrtcSignalingResponse } from "../types/MCMCMonitorRequest";
+import { WebrtcSignalingRequest, WebrtcSignalingResponse } from "../types/MCMCMonitorRequestTypes";
 
 class SignalCommunicator {
     #onConnectionCallbacks: ((connection: SignalCommunicatorConnection) => void)[] = []
