+++ service/src/types/MCMCMonitorPeerRequestTypes.ts
@@ -1,0 +1,13 @@
+import { MCMCMonitorRequest, MCMCMonitorResponse } from "./MCMCMonitorRequestTypes"
+
+export type MCMCMonitorPeerRequest = {
+    type: 'mcmcMonitorPeerRequest'
+    request: MCMCMonitorRequest
+    requestId: string
+}
+
+export type MCMCMonitorPeerResponse = {
+    type: 'mcmcMonitorPeerResponse'
+    response: MCMCMonitorResponse
+    requestId: string
+}
