+++ service/src/types/MCMCMonitorTypes.ts
@@ -1,13 +1,6 @@
-import validateObject, { isArrayOf, isBoolean, isNumber, isString, optional } from "./validateObject"
-
 export type MCMCRun = {
     runId: string
 }
-
-export const isMCMCRun = (x: any): x is MCMCRun => (
-    validateObject(x, { runId: isString })
-)
-
 
 export type MCMCChain = {
     runId: string
@@ -20,20 +13,6 @@ export type MCMCChain = {
     lastChangeTimestamp: number
 }
 
-export const isMCMCChain = (x: any): x is MCMCChain => (
-    validateObject(x, {
-        runId: isString,
-        chainId: isString,
-        variableNames: isArrayOf(isString),
-        rawHeader: optional(isString),
-        rawFooter: optional(isString),
-        variablePrefixesExcluded: optional(isArrayOf(isString)),
-        excludedInitialIterationCount: optional(isNumber),
-        lastChangeTimestamp: isNumber
-    })
-)
-
-
 export type MCMCSequence = {
     runId: string
     chainId: string
@@ -41,13 +20,3 @@ export type MCMCSequence = {
     data: number[]
     updateRequested?: boolean
 }
-
-export const isMCMCSequence = (x: any): x is MCMCSequence => (
-    validateObject(x, {
-        runId: isString,
-        chainId: isString,
-        variableName: isString,
-        data: isArrayOf(isNumber),
-        updateRequested: optional(isBoolean)
-    })
-)
