+++ service/src/types/WebsocketMessageTypes.ts
@@ -1,0 +1,4 @@
+export type WebsocketMessage = {
+    type: 'signal'
+    signal: string
+}
