+++ service/src/types/index.ts
@@ -1,0 +1,10 @@
+
+
+export * from './ConnectorHttpProxyTypes'
+export * from './MCMCMonitorPeerRequestTypes'
+export * from './MCMCMonitorRequestTypes'
+export * from './MCMCMonitorTypes'
+export * from './Typeguards'
+export * from './WebsocketMessageTypes'
+
+export const protocolVersion = '0.1.4'
