+++ src/MCMCMonitorDataManager/MCMCMonitorData.ts
@@ -1,5 +1,5 @@
 import React from 'react'
-import { MCMCChain, MCMCRun, MCMCSequence } from '../../service/src/types/MCMCMonitorTypes'
+import { MCMCChain, MCMCRun, MCMCSequence } from '../../service/src/types'
 
 export type GeneralOpts = {
     dataRefreshMode: 'auto' | 'manual'
