+++ src/networking/WebrtcConnectionToService.ts
@@ -1,6 +1,5 @@
 import SimplePeer from "simple-peer";
-import { MCMCMonitorPeerRequest, isMCMCMonitorPeerResponse } from "../../service/src/types/MCMCMonitorPeerRequest";
-import { MCMCMonitorRequest, MCMCMonitorResponse, WebrtcSignalingRequest } from "../../service/src/types/MCMCMonitorRequest";
+import { MCMCMonitorPeerRequest, MCMCMonitorRequest, MCMCMonitorResponse, WebrtcSignalingRequest, isMCMCMonitorPeerResponse } from "../../service/src/types";
 import postApiRequest from "./postApiRequest";
 
 class WebrtcConnectionToService {
