+++ test/MCMCMonitorDataManager/MCMCMonitorData.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, test } from 'vitest'
-import { MCMCChain } from '../../service/src/types/MCMCMonitorTypes'
+import { MCMCChain } from '../../service/src/types'
 import { MCMCMonitorData, SequenceStatsDict, VariableStatsDict, appendData_TEST, chainsWereUpdated_TEST, computeEffectiveWarmupIterations_TEST, detectedWarmupIterationCount, doChainUpdate_TEST, initialMCMCMonitorData, invalidateStats_TEST } from '../../src/MCMCMonitorDataManager/MCMCMonitorData'
 
 describe("Detected warmup iteration count", () => {
